mcp/PydanticAI/demo05-mcp-sse-agentic-doctor/requirements.txt
streamlit
pydantic_ai
openai
pydantic
python-dotenv
pydantic[email]
nest_asyncio
pydantic-ai-slim[mcp]
logfire
agno
yfinance